Modified 100-prime_factor.c

+++ 100-prime_factor.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-long int i, n;
+long long i, n;
 
 n = 612852475143;
 for (i = 2; i < n; i++)
